2023/langflow/pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "allainclair", email = "allainclair@gmail.com"},
]
dependencies = [
    "fastapi>=0.95.2",
    "uvicorn[standard]>=0.22.0",
    "httpx>=0.24.0",
    "langchain>=0.0.171",
    "python-dotenv>=1.0.0",
    "openai>=0.27.6",
    "tiktoken>=0.4.0",
]
requires-python = ">=3.11"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.3.1",
    "mypy>=1.3.0",
    "icecream>=2.1.3",
    "pytest-asyncio>=0.21.0",
]

[tool.pdm.scripts]
test = "pytest -ssvv"
run-dev = "pdm run uvicorn app.main:app --reload"
